requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
pikepdf==9.2.1
Pillow==10.4.0